Fix CI failures from bytecode-parity work
- clippy: drop redundant `test_` prefix on three test functions and
  remove an unnecessary `u32` cast in basicblock_clear_reuses_cpython_spare_slots_in_offset_order
- insta: regenerate nested_double_async_with snapshot to match the new
  CFG output that drops unreferenced labels after the redundant-NOP pass
- regrtest: drop `@expectedFailure` markers from test_func_args,
  test_meth_args (test_compile), test_disassemble_with,
  test_disassemble_try_finally (test_dis), and test_except_star
  (test_monitoring) which now pass

+++ crates/codegen/src/ir.rs
@@ -7208,7 +7208,7 @@ mod tests {
         for i in 0..3 {
             let mut stale = test_instr(Instruction::Nop, 35 + i);
             stale.except_handler = Some(ExceptHandlerInfo {
-                handler_block: BlockIdx::new(i as u32 + 1),
+                handler_block: BlockIdx::new(i + 1),
                 preserve_lasti: false,
             });
             test_block_push(&mut block, stale);
